feat[#27] : prettier 적용
- npm install prettier 를 통해 prettier 적용
- 관련해서 잘 되는지 체크해 보았고, .prettierrc 파일에 관련 규약을 적어놓음
