+++ FrontEnd/rebon-frontend/src/.prettierrc
@@ -1,0 +1,8 @@
+{
+  "singleQuote": true,
+  "semi": true,
+  "useTabs": false,
+  "tabWidth": 2,
+  "printWidth": 180,
+  "arrowParens": "always"
+}
